+++ kernel/tests/ui/invalid_nan.rs
@@ -1,6 +1,6 @@
 use kernel::time_temperature_curve::impl_polyline_const::ImplPolylineConst;
 
 const _INVALID_NAN: ImplPolylineConst<2> =
-    ImplPolylineConst::from_array([(0.0, f64::NAN), (1.0, 2.0)]);
+    ImplPolylineConst::from_array([(0.0, f32::NAN), (1.0, 2.0)]);
 
 fn main() {}
